requirements.txt
opencv-python
torch
torchvision
pyyaml
easydict
scikit-image
matplotlib
scipy
scikit-learn